Seperate serialisers from LogData class
This commit seperates out serialisation methods from the LogData class,
whilst making them convertable between each other. This then enforces
the concept of each class having their own responsibilities.
